pyproject.toml
[project]
name = "mixpilot"
version = "0.1.0"
description = "MixPilot — real-time audio analysis and mixing assistant for live audio environments"
authors = [
    { name = "Jay", email = "wjje5ng@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "numpy>=2.4.4",
    "pydantic-settings>=2.14.1",
    "pyloudnorm>=0.2.0",
    "python-osc>=1.10.2",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
    "uvicorn[standard]>=0.46.0",
]

[build-system]
requires = ["uv_build>=0.11.12,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-ra", "--strict-markers"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
